Fixed issues and updated migrations

+++ BurnoutTracker/Domain/Models/Entities/RepositoryApi.cs
@@ -3,7 +3,7 @@
     public class RepositoryApi
     {
         public int Id { get; set; }
-        public int SupportedRepositoryId { get; set; }
+        public long SupportedRepositoryId { get; set; }
         public SupportedRepository SupportedRepository { get; set; } = null!;
         public string Name { get; set; } = string.Empty; // e.g., "Get Contributors"
         public string Path { get; set; } = string.Empty; // e.g., "/repos/{owner}/{repo}/contributors"
